refactor(opencode): extract the delivery ledger record schema validation
The prompt-delivery ledger is a versioned JSON file that outlives the process
that wrote it, so every record read back is untrusted input until a closed set
of type guards has looked at it. That set - the record guard, the scalar and
array guards under it, and the enumerations they check membership in - changes
only when the record shape changes, which is a different rhythm from the store
that reads and writes those records.

Moved verbatim into `OpenCodePromptDeliveryLedgerRecordSchema.ts`, with no
behaviour change: the store keeps `validate:` pointing at the same function and
the file keeps its own `isOpenCodePromptDeliveryCancelled` and retention rules,
which are about transitions rather than shape.

The move is what makes room under the source-size guard for the following
commits, and the validation gains the tests it never had on the way out - the
three failure messages, the enum and scalar rejections that produce them, and
the optional fields a long-lived record accumulates.
